created openinghours hook
