+++ src/modules/map/utils.ts
@@ -33,7 +33,6 @@ import {
 } from '@atb/modules/mobility';
 import {SLIGHTLY_RAISED_MAP_PADDING} from './MapConfig';
 import turfBooleanPointInPolygon from '@turf/boolean-point-in-polygon';
-import opening_hours from 'opening_hours';
 
 export const hitboxCoveringIconOnly = {width: 1, height: 1};
 
@@ -318,31 +317,4 @@ export function getFeatureWeight(
   } else {
     return 0;
   }
-}
-
-export function getOpeningHoursInterval(oh: opening_hours, timeNow: Date) {
-  const today = new Date(timeNow);
-  today.setHours(0, 0, 0, 0);
-
-  const tomorrow = new Date(today);
-  tomorrow.setDate(today.getDate() + 1);
-
-  const openIntervalsToday = oh?.getOpenIntervals(today, tomorrow);
-
-  if (
-    !openIntervalsToday ||
-    openIntervalsToday.length === 0 ||
-    !openIntervalsToday[0]?.[0] ||
-    !openIntervalsToday[0]?.[1]
-  ) {
-    return {open: null, closing: null};
-  }
-
-  return {open: openIntervalsToday[0][0], closing: openIntervalsToday[0][1]};
-}
-
-export function isScooterOpen(oh: opening_hours, dateNow: Date) {
-  if (!oh) return false;
-
-  return oh?.getState(dateNow);
 }
